feat: update installation script for az, aws, terraform & decommision for rhel script, centralize script for linux

+++ linux/installation/developer-packages/ubuntu/tools/aws_cli.sh
@@ -1,7 +1,19 @@
 #! /bin/bash
 
-curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
-unzip awscliv2.zip
-./aws/install
-echo "==== CLEAN UP ===="
-rm -f awscliv2.zip && rm -drf aws
+ARCH=$(uname -m)
+
+if [ "$ARCH" = "aarch64" ]; then
+    curl "https://awscli.amazonaws.com/awscli-exe-linux-aarch64.zip" -o "awscliv2.zip"
+    unzip awscliv2.zip
+    sudo ./aws/install
+    echo "==== CLEAN UP ===="
+    rm -f awscliv2.zip && rm -drf aws
+fi
+
+if [ "$ARCH" = "x86_64" ]; then
+    curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
+    unzip awscliv2.zip
+    sudo ./aws/install
+    echo "==== CLEAN UP ===="
+    rm -f awscliv2.zip && rm -drf aws
+fi
